% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_functions.R
\name{abc_moran_deme2}
\alias{abc_moran_deme2}
\title{Parallelized abc_moran_deme}
\usage{
abc_moran_deme2(
  nsims,
  t,
  priors,
  x.max = 100,
  y.max = 100,
  spatial = NULL,
  eqpop = FALSE,
  eqmig = TRUE,
  outgens = NULL,
  parallel = TRUE,
  n_cores = NULL
)
}
\description{
Parallelized abc_moran_deme
}
