% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{birth_death_process}
\alias{birth_death_process}
\title{Conduct a single birth death process}
\usage{
birth_death_process(x, params)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{params}{params object}
}
\value{
numeric matrix, updated matrix
}
\description{
Conduct a single birth death process
}
