% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{calculate_species_probs}
\alias{calculate_species_probs}
\title{Calculate species probabilities}
\usage{
calculate_species_probs(x, com_index, params)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{com_index}{Int, row index for community to sample}

\item{params}{Numeric matrix, matrix of selection coefficients}
}
\value{
int, index of species for birth
}
\description{
Calculate species probabilities
}
