% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic_functions.R
\name{d_gamma}
\alias{d_gamma}
\title{Calculate True Gamma Diversity}
\usage{
d_gamma(x, q = 1, weight = FALSE)
}
\arguments{
\item{x}{Numeric matrix. A meta-community abundance matrix (species x site)  where columns
are species, rows are communities and cell ij is the count of species j in community i.}

\item{q}{Numeric. Order of the diversity measure.}

\item{weight}{Logical indicating if communities should be given different weights.
Following Jost 2007, this is only valid when q=1 so if set to `TRUE`, q is 
set to 1 and weights are set according to the sizes of individual communities.}
}
\value{
Numeric, the true alpha diversity
}
\description{
Function to calculate the true gamma diversity of a metacommunity
}
\details{
This function calculates true gamma diversity using numbers equivalents
as detailed in Jost 2008.
}
\examples{
x <- rand_meta(10, 6, 50)
d_gamma(x, q=1)

}
