% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_functions.R
\name{is.params}
\alias{is.params}
\title{Check if object is Params class}
\usage{
is.params(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical
}
\description{
Checks to see if an object is of class params
}
\examples{

test <- make_params(5,5)
is.params(test)

}
