% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_functions.R
\name{is.priors}
\alias{is.priors}
\title{Check if object is priors class}
\usage{
is.priors(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical
}
\description{
Checks to see if an object is of class priors
}
\examples{

test <- make_priors(5,5)
is.priors(test)

}
