% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_functions.R
\name{make_params}
\alias{make_params}
\title{Create params object.}
\usage{
make_params(n.spec, n.site)
}
\arguments{
\item{n.spec}{Numeric, the number of species in the metacommunity.}

\item{n.site}{Numeric, the number of communities in the metacommunity.}
}
\value{
Returns an empty object of S3 class \emph{params}.
}
\description{
Initializes an object of class \emph{params}.
}
\details{
This function is used to initialize a \emph{params} object.
  A \emph{params} object is a list with 3 named elements: s, fd, and mig.
  s is a matrix of selection coefficients where columns correspond to species
  and rows correspond to sites (or communities). fd is a vector containing
  the frequency dependence parameters for each species. mig is a list of
  migration matrices, where cell ij is the probability for each species to
  migrate from site j to site i.
}
\examples{
make_params(5,5)

}
