% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_functions.R
\name{make_selfromenv}
\alias{make_selfromenv}
\title{Create Selection Matrices from Environmnetal data}
\usage{
make_selfromenv(
  n.spec,
  env,
  env.maxmin,
  var.maxmin,
  meanvals = NULL,
  covar = NULL
)
}
\arguments{
\item{n.spec}{Integer, number of species in metacommunity}

\item{env}{Matrix or Data-Frame with environmental values measured
from communities.}

\item{env.maxmin}{List with length equal to number of environmental measures.
Each element in the list is a Numeric vector with two elements, the minimum
and maximum values for each environmental variable.}

\item{var.maxmin}{List with length equal to number of environmental measures.
Each element in the list is a Numeric vector with two elemens, the minimum
and maximum values that the variance can be for each environmental variable.}

\item{meanvals}{Matrix, user inputed value of means for each species. This is
an m x n matrix where columns are environmental variables and rows are species.
Each cell is the value of the environmental variable at which the species is
found at the highest frequency.}

\item{covar}{User inputed list of covariance matrices.}
}
\value{
A matrix of selection coefficients where columns are species, rows are
communities, and cell ij is the selection coefficient for species j in community i.
}
\description{
This function creates random selection matrices by generating
resource-utlization niches for N number of species according
to inputted environmental data for communities.
}
\details{
This function creates a selection matrix by randomly generating
multivariate normal Resource Utilization niches and matching those to generated
environmental variables for each community using \code{\link{make_spatialenv}}.
First, values for each environmental variable are calculated  for each
species using a random uniform distribution. This value is considered the peak
for the species niche for that environmental variable  (i.e. the value at which
the species is found at the highest frequency). Each species is then generated
a random variance for each environmental variable according to the var.maxmin
argument. At this point, this function assumes that the covariance for each
environmental variable is 0. Alternatively, the user can input means and covariances
by hand. From these values, the multivariate density for each species is calculated.
Species coefficients are then generated by dividing the density of each species in each
community by the density of the species with the max density.
}
\examples{

testmaxmin <- list(pH = c(6, 8), eC = c(.01, .15), temp = c(70, 85))
testenv <- make_spatialenv(n.sites = 10, n.meas = 3, env.maxmin = testmaxmin, autocor = TRUE)
testvar <- list(pH = c(2,5), eC = c(.05, .1), Temp = c(5, 8))
make_selfromenv(n.spec = 5, env = testenv$env.vals, env.maxmin = testmaxmin, var.maxmin = testvar)

}
