% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic_functions.R
\name{morisitas}
\alias{morisitas}
\title{Calculate Morisita's Index}
\usage{
morisitas(x)
}
\arguments{
\item{x}{A metacommunity matrix with species as columns and sites as rows}
}
\description{
Calculate Morisita's index for a metacommunity
}
\details{
This function calculates Morisita's index following (TODO: add reference)
}
\examples{
meta <- rand_meta(10, 15, 100)
morisitas(meta)

}
