% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{normalize_meta}
\alias{normalize_meta}
\title{Normalize Metacommunity}
\usage{
normalize_meta(x)
}
\arguments{
\item{x}{A metacommunity matrix (site x species) with cells as species counts}
}
\description{
Normalizes a metacommunity matrix using community sizes
}
\details{
Converts a metacommunity with cell counts to cell frequencies based on 
community sizes (i.e. species frequency in community i)
}
