% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_functions.R
\name{print.priors}
\alias{print.priors}
\title{S3 print method for priors}
\usage{
\method{print}{priors}(x)
}
\arguments{
\item{x}{Object of class priors}
}
\value{
Prints out to screen
}
\description{
Sets up the print methods for class priors
}
\examples{
priors <- make_priors(5, 5)
print(priors)

}
