% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{rand_meta}
\alias{rand_meta}
\title{Random metacommunity matrix.}
\usage{
rand_meta(N, S, J, min.spec = 2)
}
\arguments{
\item{N}{Numeric, the number of communities.}

\item{S}{Numeric, the number of species.}

\item{J}{Numeric vector, the number of individuals within each community. If
length = 1 gives all communities the same number of individuals.}

\item{min.spec}{Numeric, the minimum number of species within communities.}
}
\value{
Returns a numeric matrix where rows are communities, columns are species,
and cell ij is the count of species j in community i.
}
\description{
Creates a random metacommunity matrix where cell ij is the number of species j in
community i.
}
\details{
This function creates a random metacommunity matrix in which species
absolute abundances \(counts\) within communities are recorded.
}
\examples{
rand_meta(N = 5, S = 5, J = c(100, 100, 200, 40, 300), min.spec = 2)
rand_meta(N = 5, S = 5, J = 500)

\dontrun{
rand_meta(N = 5, S = 5, J = c(10, 10))
}

}
