% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{random_points}
\alias{random_points}
\title{Random Cartesian Points}
\usage{
random_points(n.sites, x.max, y.max)
}
\arguments{
\item{n.sites}{Numeric, the number of sites to generate}

\item{x.max, y.max}{Numeric, the maximum distance in x or y directions that sites can
be placed.}
}
\value{
Returns a two-column matrix of coordinate points.
}
\description{
Creates a two-column matrix where the first column is x-values and the second column
is y-values and each row represents a point. Can be used to randomly generate a
spatially explicit list of sites.
}
\examples{
random_points(20, 100, 100)

}
