% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{samp_com_for_birth}
\alias{samp_com_for_birth}
\title{Select community for birth}
\usage{
samp_com_for_birth(x)
}
\arguments{
\item{x}{Numeric matrix, species x site}
}
\value{
int, index of community where birth is to occur
}
\description{
Select community for birth
}
