% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{samp_com_for_death}
\alias{samp_com_for_death}
\title{Select Community for death}
\usage{
samp_com_for_death(x, species_index, com_index, params)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{species_index}{Int, row index for community to sample}

\item{com_index}{Int, community of birth species}

\item{params}{Params object}
}
\value{
int, index of species for birth
}
\description{
Select Community for death
}
