% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{samp_species_for_birth}
\alias{samp_species_for_birth}
\title{Select species for birth}
\usage{
samp_species_for_birth(x, com_index, selection_matrix)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{com_index}{Int, row index for community to sample}

\item{selection_matrix}{Numeric matrix, matrix of selection coefficients}
}
\value{
int, index of species for birth
}
\description{
Select species for birth
}
