% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{samp_species_for_death}
\alias{samp_species_for_death}
\title{Select species for death}
\usage{
samp_species_for_death(x, com_index)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{com_index}{Int, community index for death community}
}
\value{
int, index of species for birth
}
\description{
Select species for death
}
