% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_functions.R
\name{set_fd}
\alias{set_fd}
\title{Random frequency dependence parameters.}
\usage{
set_fd(paramfile, input1, input2)
}
\arguments{
\item{paramfile}{\emph{params} object initialized using \code{\link{make_params}}.}

\item{input1, input2}{Numeric, parameters for uniform distribution.}
}
\value{
Returns a numeric matrix of selection coefficients where cell ij is the selection
coefficient for species j in community i.
}
\description{
Function to create random vector of frequency dependence parameters.
}
\details{
This function randomly generates frequency dependence parameters by sampling
from a random uniform distribution where input1 is the minimum and input2 is the
maximum.
}
\examples{
paramfile <- make_params(5,5)
paramfile$fd <- set_fd(paramfile, -1, 0)

}
