% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_functions.R
\name{set_sel}
\alias{set_sel}
\title{Random selection matrix.}
\usage{
set_sel(paramfile, distr, input1, input2)
}
\arguments{
\item{paramfile}{\emph{params} object initialized using \code{\link{make_params}}.}

\item{distr}{Character, one of 'uniform', 'normal', 'gamma', or 'beta'.}

\item{input1, input2}{Numeric, parameters for distribution being used. See details.}
}
\value{
Returns a matrix of selection coefficients where cell ij is the selection
coefficient for species j in community i.
}
\description{
Function to create random selection matrix using uniform, normal, or gamma distributions.
}
\details{
This function generates a random selection matrix according to a specified
  distribution. The specified distribution can be one of 'uniform', 'normal', or 'gamma.'
  If 'uniform' input1 is the minimum value and input2 is the maximum value. If 'normal'
  input1 is the mean and input2 is the standard deviation. If 'gamma' input1 is the
  shape and input2 is the scale parameters. if 'beta' input1 is the first shape
  parameter and input2 is the second shape parameter
}
\examples{

paramfile <- make_params(5,5)
paramfile$s <- set_sel(paramfile, 'normal', .5, .2)

## For neutral model ##
paramfile <- make_params(5,5)
paramfile$s <- set_sel(paramfile, 'uniform', 1,1)

}
