% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic_functions.R
\name{ssid2beta}
\alias{ssid2beta}
\title{Estimate beta-diversity using sum of squared interpoint dissimilarities}
\usage{
ssid2beta(x, method = "bray")
}
\arguments{
\item{x}{metacommunity matrix}

\item{method}{type of dissimilarity metric used; one of the methods in vegdist()}
}
\description{
Calculates the sum of squared interpoint dissimilarities of a metacommunity
matrix using the formula in Anderson et al. (2011) in Ecology Letters
}
\details{
Estimates beta-diversity of a metacommunty matrix by using the sum of
squared interpoint dissimilarities: 1/(N(N-1))*sum(dij^2) where dij is the distance
between community i and j.
}
