% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic_functions.R
\name{turnover}
\alias{turnover}
\title{Calculate Species Turnover}
\usage{
turnover(x)
}
\arguments{
\item{x}{A metacommunity matrix with species as columns and sites as rows}
}
\description{
Calculate the species turnover of a metacommunity matrix
}
\details{
This function calculates species turnover following (TODO: add reference)
}
\examples{
meta <- rand_meta(10, 15, 100)
turnover(meta)

}
