% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{update_meta}
\alias{update_meta}
\title{Update metacommunity matrix}
\usage{
update_meta(x, com_index, b_species_index, d_species_index)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{com_index}{Int, community index for community where death occurs}

\item{b_species_index}{Int, index for species for birth}

\item{d_species_index}{Int, index for species for death}
}
\value{
numeric matrix, updated matrix
}
\description{
Update metacommunity matrix
}
