% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_functions.R
\name{abc_moran_deme}
\alias{abc_moran_deme}
\title{Run Multiple Moran Community Simulations With Priors.}
\usage{
abc_moran_deme(
  nsims,
  t,
  priors,
  x.max = 100,
  y.max = 100,
  spatial = NULL,
  eqpop = FALSE,
  eqmig = TRUE,
  change_params = TRUE,
  prop_new = 0.05,
  diff_sd = 0.05,
  parallel = TRUE,
  n_cores = NULL
)
}
\arguments{
\item{nsims}{Numeric, The number of simulations to run.}

\item{t}{Numeric, the number of generations to run for each simulation.}

\item{priors}{Priors object specifying the priors to use for simulations. See \code{\link{make_priors}}.}

\item{x.max, y.max}{Numeric, maximum spatial extent in x or y direction for site placement if randomly
generating spatial data.}

\item{spatial}{Two column Numeric matrix or Distance object specifying spatial arrangement of communities.}

\item{eqpop}{Logical, if TRUE all community sizes are the same within a simulation, if FALSE
each community is randomly given a community size each simulation.}

\item{eqmig}{Logical, if TRUE all species will have the same migration matrix, if FALSE
each species will recieve it's own migration matrix each simulation according to priors file settings}

\item{change_params}{Boolean, if true speciate "evolves" the input params object}

\item{prop_new}{Numeric, the proportion of individuals in the community that should be replaced
when speciation occurs}

\item{diff_sd}{Numeric, the standard deviation used when "evolving" selection and fd coefficients}

\item{parallel}{Boolean, should the simulations be run in parallel}

\item{n_cores}{the number of cores to use. If Null sets the number of cores
to 2 less than the computer contains}
}
\value{
If all communities have the same size, then returns a list with a list of final metacommunity matrices
for each simulation, a list of selection matrices used for each simulation, a matrix with frequency dependence
parameters for each species (columns are species and rows are simulation runs), and a dataframe with the
parameters used for each simulation (each row corresponds to one simulation). If communities are allowed to have
different sizes, then the list also contains a matrix containing the community sizes for each community in each
simulation.
}
\description{
This function runs multiple simulations of the Moran Community model by randomly generating parameters
according to user input prior distributions for use in downstream ABC analyses.
}
\details{
This function is used to run the Moran Community model simulation multiple times in preparation
for Approximate Bayesian Analysis (ABC). Users specify prior distributions for community size,
selection coefficients, frequency dependence, and migration parameters in a \emph{priors} object. From
this information, parameter values are randomly drawn according to the prior distributions and simulations
are run from these parameter values.
}
\examples{
\dontrun{
testpriors <- make_priors(5, 5)
xy <- random_points(5, 100, 100)

testpriors$jdist <- 1
testpriors$jparams <- c(100, 200)
testpriors$seldist <- 2
testpriors$selparams <- c(.5, .1)
testpriors$fddist <- 1
testpriors$fdparams <- c(-.5, -.2)
testpriors$migdist <- 1
testpriors$migdistparams <- c(10, 50)
testpriors$migprobdist <- 1
testpriors$migprobparams <- c(.1, .2)

abc_moran_deme(5, 5, testpriors, eqpop = FALSE, spatial = xy, parallel = FALSE)
}

}
