% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{ecosumstats}
\alias{ecosumstats}
\title{Calculate Summary Statistics}
\usage{
ecosumstats(y)
}
\arguments{
\item{y}{Either a metacommunity matrix or a list of metacommunity matrices}
}
\value{
Either a vector of summary statistics for a single metacommunity or a dataframe where columns
are different summary statistics and rows are metacommunities.
}
\description{
Calculates summary statistics on metacommunities to be used in ABC analyses.
}
\details{
Calculates the mean, median, and standard deviations for shannon diversity index,
simpson diversity index, Pielou's evenness, richness, beta diversity (whittakers and Sorensons),
and the number of sites occupied for species for communities in a metacommunity matrix. It also
calculats the mean and median C-score and the v-ratio.
}
\examples{
meta <- CommSimABC::rand_meta(20, 30, 500)
ecosumstats(meta)

metas <- replicate(20, rand_meta(20, 30, 500), simplify = FALSE)
ecosumstats(metas)

}
