% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistic_functions.R
\name{evenness}
\alias{evenness}
\title{Calculate Community Evenness.}
\usage{
evenness(x)
}
\arguments{
\item{x}{A metacommunity matrix (columns = species, rows = communities, cell values = counts).}
}
\value{
A vector of evenness values for each community.
}
\description{
Calculates Pielou's evenness of a metacommunity.
}
\examples{
meta <- rand_meta(20, 30, 500)
evenness(meta)

}
