% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_functions.R
\name{run_single_sim}
\alias{run_single_sim}
\title{Create params and run a single moran_deme simulation}
\usage{
run_single_sim(
  t,
  priors,
  x.max = 100,
  y.max = 100,
  spatial = NULL,
  eqpop = FALSE,
  eqmig = TRUE,
  output = FALSE,
  change_params = FALSE,
  prop_new = 0,
  diff_sd = 0.05
)
}
\arguments{
\item{t}{Numeric, the number of generations to run for each simulation.}

\item{priors}{Priors object specifying the priors to use for the simulation See \code{\link{make_priors}}.}

\item{x.max, y.max}{Numeric, maximum spatial extent in x or y direction for site placement if randomly
generating spatial data.}

\item{spatial}{Two column Numeric matrix or Distance object specifying spatial arrangement of communities.}

\item{eqpop}{Logical, if TRUE all community sizes are the same within a simulation, if FALSE
each community is randomly given a community size each simulation.}

\item{eqmig}{Logical, if TRUE all species will have the same migration matrix, if FALSE
each species will receive it's own migration matrix}

\item{output}{Logical, if True outputs progress bar.}

\item{change_params}{Logical, if true speciate function "evolves" the input params file}

\item{prop_new}{Numeric, the proportion of individuals in the community that should be replaced
when speciation occurs}

\item{diff_sd}{Numeric, the standard deviation used when "evolving" selection and fd coefficients}
}
\value{
simrun object with simulation results
}
\description{
This functions takes a priors object and uses it to create a 
parameters object that is used in a moran_deme simulation.
}
