% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{samp_species_for_birth}
\alias{samp_species_for_birth}
\title{Select species for birth}
\usage{
samp_species_for_birth(x, probs)
}
\arguments{
\item{x}{Numeric matrix, species x site}

\item{probs}{Numeric vector, vector of species probabilities}
}
\value{
int, index of species for birth
}
\description{
Select species for birth
}
