% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{speciate}
\alias{speciate}
\title{Create new species}
\usage{
speciate(x, params, prop_new = 0.1, diff_sd = 0.1)
}
\arguments{
\item{x}{Numeric matrix, species x site matrix}

\item{params}{S3 class params object}

\item{prop_new}{float, the proportion of the community size that the new species
should have}

\item{diff_sd}{float, the standard deviation to use when evolving the selection
coefficients and frequency dependence parameters. These values are evolved by
randomly selecting a number from a normal distribution with mean 0 and standard
deviation per this parameter.}
}
\value{
Returns a new site x species matrix with all column sums greater than
zero and a new params object with updated parameters for the new species
}
\description{
This function randomly creates new species in a site x species matrix if 
there are species (columns) with community sums equal to 0.
}
\details{
This function takes a site x species matrix and a params object and 
checks to see if there are any empty species (column) vectors. If there are
it randomly selects a single community (row) and adds individuals into the 
cell of the empty column up to the frequency specified in prop_new. To keep the 
community size the same, individuals are randomly subtracted from other non-zero
species in the community. Additionally, the selection coefficients and frequency
dependence parameters are modified from the extinct species parameters by 
adding values from a normal distribution centered at 0.
}
